chore: relocate examples.
