+++ examples/internal/helloworld/helloworld.proto
