+++ examples/postgres-query/main.go
