--- instrumentation/hypertrace/github.com/gin-gonic/hypergin/examples/go.mod
@@ -1,10 +1,0 @@
-module github.com/hypertrace/goagent/instrumentation/hypertrace/github.com/gin-gonic/gypergin/examples/main
-
-go 1.15
-
-replace github.com/hypertrace/goagent => ../../../../../../
-
-require (
-	github.com/gin-gonic/gin v1.7.2
-	github.com/hypertrace/goagent v0.2.1
-)
