--- instrumentation/hypertrace/github.com/gorilla/hypermux/examples/server/go.mod
@@ -1,10 +1,0 @@
-module github.com/hypertrace/goagent/instrumentation/hypertrace/github.com/gorilla/hypermux/examples/server
-
-go 1.15
-
-replace github.com/hypertrace/goagent => ../../../../../../../
-
-require (
-	github.com/gorilla/mux v1.8.0
-	github.com/hypertrace/goagent v0.0.0-00010101000000-000000000000
-)
