+++ instrumentation/opentelemetry/google.golang.org/hypergrpc/internal/helloworld/helloworld.proto
