--- instrumentation/opentelemetry/net/hyperhttp/examples/client/main.go
@@ -1,61 +1,0 @@
-// +build ignore
-
-package main
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"log"
-	"net/http"
-
-	"github.com/hypertrace/goagent/config"
-	"github.com/hypertrace/goagent/instrumentation/opentelemetry"
-	"github.com/hypertrace/goagent/instrumentation/opentelemetry/net/hyperhttp"
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
-)
-
-type message struct {
-	Content string `json:"message"`
-}
-
-func main() {
-	cfg := config.Load()
-	cfg.ServiceName = config.String("http-client")
-
-	closer := opentelemetry.Init(cfg)
-	defer closer()
-
-	client := http.Client{
-		Transport: otelhttp.NewTransport(
-			hyperhttp.WrapTransport(http.DefaultTransport),
-		),
-	}
-
-	req, err := http.NewRequest("GET", "http://localhost:8081/foo", bytes.NewBufferString(`{"name":"Dave"}`))
-	req.Header.Set("Content-Type", "application/json")
-	if err != nil {
-		log.Fatalf("failed to create the request: %v", err)
-	}
-
-	res, err := client.Do(req)
-	if err != nil {
-		log.Fatalf("failed to perform the request: %v", err)
-	}
-
-	resBody, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		log.Fatalf("failed to read the response body: %v", err)
-	}
-	defer res.Body.Close()
-
-	m := &message{}
-	err = json.Unmarshal(resBody, m)
-	if err != nil {
-		log.Fatalf("failed to unmarshal the response body: %v", err)
-		return
-	}
-
-	fmt.Println(m.Content)
-}
