--- instrumentation/opentelemetry/net/hyperhttp/examples/server/main.go
@@ -1,63 +1,0 @@
-// +build ignore
-
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"log"
-	"net/http"
-	"time"
-
-	"github.com/gorilla/mux"
-	"github.com/hypertrace/goagent/config"
-	"github.com/hypertrace/goagent/instrumentation/opentelemetry"
-	"github.com/hypertrace/goagent/instrumentation/opentelemetry/net/hyperhttp"
-	sdkhttp "github.com/hypertrace/goagent/sdk/instrumentation/net/http"
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
-)
-
-func main() {
-	cfg := config.Load()
-	cfg.ServiceName = config.String("http-server")
-
-	closer := opentelemetry.Init(cfg)
-	defer closer()
-
-	r := mux.NewRouter()
-	r.Handle("/foo", otelhttp.NewHandler(
-		hyperhttp.WrapHandler(
-			http.HandlerFunc(fooHandler),
-			&sdkhttp.Options{},
-		),
-		"/foo",
-	))
-	log.Fatal(http.ListenAndServe(":8081", r))
-}
-
-type person struct {
-	Name string `json:"name"`
-}
-
-func fooHandler(w http.ResponseWriter, r *http.Request) {
-	sBody, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	defer r.Body.Close()
-
-	p := &person{}
-	err = json.Unmarshal(sBody, p)
-	if err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	<-time.After(300 * time.Millisecond)
-
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte(fmt.Sprintf("{\"message\": \"Hello %s\"}", p.Name)))
-}
