+++ sdk/instrumentation/google.golang.org/grpc/common_test.go
@@ -8,7 +8,7 @@ import (
 	"net"
 	reflect "reflect"
 
-	"github.com/hypertrace/goagent/instrumentation/hypertrace/google.golang.org/hypergrpc/examples/helloworld"
+	"github.com/hypertrace/goagent/sdk/instrumentation/google.golang.org/grpc/internal/helloworld"
 	"github.com/hypertrace/goagent/sdk/internal/mock"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
